backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
pgvector==0.3.3
redis==5.0.8
boto3==1.35.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
httpx==0.27.2
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
pypdf==5.0.1
sse-starlette==2.1.3
numpy==2.1.1
faiss-cpu==1.9.0.post1
